requirements.txt
fastapi
uvicorn
pyodbc
pandas
requests
python-dotenv
Office365-REST-Python-Client
requests_ntlm
pillow
python-magic